add POST

+++ professions/src/main/java/br/com/ufsm/professions/repository/TitleRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ufsm.professions.model.Title;
 
-public interface TitleRepository extends JpaRepository<Title, Integer> {
+public interface TitleRepository extends JpaRepository<Title, Long> {
 	
 	Title findByName (String name);
 }
